+++ MyCollection/Pages/Settings/Signatures/Delete.cshtml.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
@@ -10,13 +8,16 @@ using MyCollection.Models;
 
 namespace MyCollection.Pages.Signatures
 {
+    [Authorize(Roles = "Basic")]
     public class DeleteModel : PageModel
     {
-        private readonly MyCollection.Data.MyCollectionContext _context;
+        private readonly MyCollectionContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public DeleteModel(MyCollection.Data.MyCollectionContext context)
+        public DeleteModel(MyCollectionContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         [BindProperty]
@@ -29,7 +30,9 @@ namespace MyCollection.Pages.Signatures
                 return NotFound();
             }
 
-            var signature = await _context.Signatures.FirstOrDefaultAsync(m => m.Id == id);
+            var signature = await _context.Signatures
+                .Include(s=>s.Person)
+                .FirstOrDefaultAsync(m => m.Id == id);
 
             if (signature == null)
             {
@@ -37,6 +40,11 @@ namespace MyCollection.Pages.Signatures
             }
             else 
             {
+                var user = await _userManager.GetUserAsync(User);
+                if (user == null || signature.User != user)
+                {
+                    return RedirectToPage("/AccessDenied");
+                }
                 Signature = signature;
             }
             return Page();
@@ -52,6 +60,11 @@ namespace MyCollection.Pages.Signatures
 
             if (signature != null)
             {
+                var user = await _userManager.GetUserAsync(User);
+                if (user == null || signature.User != user)
+                {
+                    return RedirectToPage("/AccessDenied");
+                }
                 Signature = signature;
                 _context.Signatures.Remove(Signature);
                 await _context.SaveChangesAsync();
